precollection/requirements.txt
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.66.0
yt-dlp>=2024.10.7
faster-whisper>=1.1.0
jieba>=0.42.1
openpyxl>=3.1.2
pyyaml>=6.0.1
scikit-learn>=1.3.0
numpy>=1.24.0
